// client/src/module_bindings/types.ts
export const MarketPrice = __t.object("MarketPrice", {
  id: __t.u64(),
  crop: __t.string(),
  mandi: __t.string(),
  district: __t.string(),
  pricePerQuintalPaise: __t.u64(),
  mspPricePaise: __t.u64(),
  date: __t.string(),
});
export type MarketPrice = __Infer<typeof MarketPrice>;